# part-one/api/app/Models/Event.php
<?php

declare(strict_types=1);

namespace App\Models;

use Database\Factories\EventFactory;
use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;
use Illuminate\Support\Carbon;

/**
 * @property int $id
 * @property string $name
 * @property Carbon $date
 * @property int $registrations
 * @property Carbon $created_at
 * @property Carbon $updated_at
 */
class Event extends Model
{
    /** @use HasFactory<EventFactory> */
    use HasFactory;

    protected $fillable = ['name', 'date'];

    protected $casts = [
        'date' => 'date',
        'registrations' => 'integer',
    ];
}